refactor: move cloudinary function into seperate file
