+++ src/components/EditProfileModal/EditProfileModal.css
@@ -43,6 +43,7 @@
   height: 6rem;
   border-radius: 50%;
   object-fit: cover;
+  object-position: top;
 }
 
 .avatar {
