+++ src/components/EditProfileModal/EditProfileModal.jsx
@@ -3,6 +3,7 @@ import { toast } from "react-toastify";
 import { useSelector, useDispatch } from "react-redux";
 import { editUser } from "../../redux/asyncThunk";
 import { updateUser } from "../../redux/slices/";
+import { saveImageToCloudindary } from "../../services/";
 import { AiFillCamera, AiOutlineClose } from "react-icons/ai";
 import "./EditProfileModal.css";
 
@@ -40,47 +41,6 @@ const EditProfileModal = ({ setShowEditModal, showEditModal, userProfile }) => {
         });
       }
     };
-  };
-
-  const saveImageToCloudindary = async (
-    image,
-    saveData,
-    inputData,
-    type,
-    from = "profile"
-  ) => {
-    try {
-      const data = new FormData();
-      data.append("file", image);
-      data.append("upload_preset", "oxvrurru");
-
-      const requestOptions = {
-        method: "POST",
-        body: data,
-      };
-
-      await fetch(
-        `https://api.cloudinary.com/v1_1/abhijeetscloud/${type}/upload`,
-        requestOptions
-      )
-        .then((response) => {
-          return response.json();
-        })
-        .then((data) => {
-          let updatedData;
-          if (from === "post") {
-            updatedData = { ...inputData, img: data.url };
-          } else {
-            updatedData = { ...inputData, avatarUrl: data.secure_url };
-          }
-          saveData(updatedData);
-        })
-        .catch((error) => {
-          toast.error(error);
-        });
-    } catch (error) {
-      toast.error(error);
-    }
   };
 
   const saveEditedData = async (data) => {
